update OTP in passslip, prf, and oneUI
