+++ apps/portal/src/pages/[id]/index.tsx
@@ -133,18 +133,18 @@ export default function Dashboard({
   );
 }
 
-// export const getServerSideProps: GetServerSideProps = async (
-//   context: GetServerSidePropsContext
-// ) => {
-//   const userDetails = employeeDummy;
+export const getServerSideProps: GetServerSideProps = async (
+  context: GetServerSidePropsContext
+) => {
+  const userDetails = employeeDummy;
 
-//   return { props: { userDetails } };
-// };
+  return { props: { userDetails } };
+};
 
-export const getServerSideProps: GetServerSideProps = withCookieSession(
-  async (context: GetServerSidePropsContext) => {
-    const userDetails = getUserDetails();
-    // console.log(userDetails);
-    return { props: { userDetails } };
-  }
-);
+// export const getServerSideProps: GetServerSideProps = withCookieSession(
+//   async (context: GetServerSidePropsContext) => {
+//     const userDetails = getUserDetails();
+//     // console.log(userDetails);
+//     return { props: { userDetails } };
+//   }
+// );
