+++ libs/oneui/src/components/OtpModal/index.ts
@@ -1,0 +1,1 @@
+export * from './OtpModal';
