+++ libs/oneui/src/index.ts
@@ -19,3 +19,4 @@ export * from './components/Tables/DataTable';
 export * from './components/Spinner/LoadingSpinner';
 export * from './components/ToastNotification/ToastNotification';
 export * from './components/AlertNotification/AlertNotification';
+export * from './components/OtpModal';
